pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xswl-YPack"
version = "0.2.0"
description = "A lightweight Windows packaging tool — YAML to installer script converter (NSIS, WIX, Inno Setup)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Wang-Jianwei"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
validation = ["jsonschema>=4.0"]
web = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "jsonschema>=4.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "jsonschema>=4.0",
    "flask>=3.0",
    "flask-cors>=4.0",
]

[tool.setuptools.packages.find]
include = ["ypack*"]

[project.scripts]
xswl-ypack = "ypack.cli:main"
xswl-ypack-web = "ypack.web.server:main"

[project.urls]
Repository = "https://github.com/Wang-Jianwei/xswl-YPack"

# -----------------------------------------------------------------------
# Tool configuration
# -----------------------------------------------------------------------

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
